requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dominate==2.8.0
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
visitor==0.1.3
Werkzeug==2.3.7
WTForms==3.0.1
